+++ workspaces/features-demo/src/demo-pieces/annotation-octave-shift.ts
@@ -25,7 +25,7 @@ export function createAnnotationOctaveShiftDemo() {
         .addNote(0, "G3", "4n")
         .addNote(0, "G3", "4n")
         .addNote(0, "G3", "4n")
-        
+
         .addMeasure()
         .addNote(0, "G3", "4n").addAnnotation("loco")
         .addNote(0, "G3", "4n")
@@ -34,13 +34,13 @@ export function createAnnotationOctaveShiftDemo() {
         .endRow()
 
         .addMeasure()
-        .addNote(0, "G3", "4n").addAnnotation("8va").addExtension(ext => ext.notes("4n", 2))
+        .addNote(0, "G3", "4n").addSpan("8va", span => span.notes("4n", 2))
         .addNote(0, "G3", "4n")
         .addNote(0, "G3", "4n")
         .addNote(0, "G3", "4n")
 
         .addMeasure()
-        .addNote(0, "G3", "4n").addAnnotation("8vb").addExtension(ext => ext.notes("4n", 2))
+        .addNote(0, "G3", "4n").addSpan("8vb", span => span.notes("4n", 2))
         .addNote(0, "G3", "4n")
         .addNote(0, "G3", "4n")
         .addNote(0, "G3", "4n")