+++ CHANGELOG.md
@@ -27,6 +27,7 @@
     addAnnotation(annotationKind, annotationGroup, annotationOptions)
 - Removed function addNavigation (-> addAnnotation) and enum Navigation (-> enum AnnotationKind).
 - Renamed AnnotationOptions fields "repeatCount" -> "playCount" and "endingPassages" -> "playNumbers".
+- Replaced addAnnotation().addExtension() with addSpan().
 
 ### Added
 - "color" field to LyricsOptions.